Added opening streams

+++ spec/docx/document_spec.rb
@@ -1,18 +1,13 @@
-# coding: utf-8
 require 'docx'
 require 'tempfile'
 
 describe Docx::Document do
   before(:all) do
-    @fixtures_path = "spec/fixtures"
+    @fixtures_path = 'spec/fixtures'
     @formatting_line_count = 12 # number of lines the formatting.docx file has
   end
 
-  describe 'reading' do
-    before do
-      @doc = Docx::Document.open(@fixtures_path + '/basic.docx')
-    end
-
+  shared_examples_for 'reading' do
     it 'should read the document' do
       expect(@doc.paragraphs.size).to eq(2)
       expect(@doc.paragraphs.first.text).to eq('hello')
@@ -53,12 +48,31 @@ describe Docx::Document do
     end
   end
 
+  describe 'reading' do
+    context 'using normal file' do
+      before do
+        @doc = Docx::Document.open(@fixtures_path + '/basic.docx')
+      end
+
+      it_behaves_like 'reading'
+    end
+
+    context 'using stream' do
+      before do
+        stream = File.binread(@fixtures_path + '/basic.docx')
+        @doc = Docx::Document.open_buffer(stream)
+      end
+
+      it_behaves_like 'reading'
+    end
+  end
+
   describe 'read tables' do
     before do
       @doc = Docx::Document.open(@fixtures_path + '/tables.docx')
     end
 
-    it "should have tables with rows and cells" do
+    it 'should have tables with rows and cells' do
       expect(@doc.tables.count).to eq 2
       @doc.tables.each do |table|
         expect(table).to be_an_instance_of(Docx::Elements::Containers::Table)
@@ -71,7 +85,7 @@ describe Docx::Document do
       end
     end
 
-    it "should have tables with columns and cells" do
+    it 'should have tables with columns and cells' do
       @doc.tables.each do |table|
         table.columns.each do |column|
           expect(column).to be_an_instance_of(Docx::Elements::Containers::TableColumn)
@@ -82,23 +96,23 @@ describe Docx::Document do
       end
     end
 
-    it "should have proper count" do
+    it 'should have proper count' do
       expect(@doc.tables[0].row_count).to eq 171
       expect(@doc.tables[1].row_count).to eq 2
       expect(@doc.tables[0].column_count).to eq 2
       expect(@doc.tables[1].column_count).to eq 2
     end
 
-    it "should have tables with proper text" do
-      expect(@doc.tables[0].rows[0].cells[0].text).to eq "ENGLISH"
-      expect(@doc.tables[0].rows[0].cells[1].text).to eq "FRANÇAIS"
-      expect(@doc.tables[1].rows[0].cells[0].text).to eq "Second table"
-      expect(@doc.tables[1].rows[0].cells[1].text).to eq "Second tableau"
-      expect(@doc.tables[0].columns[0].cells[5].text).to eq "aphids"
-      expect(@doc.tables[0].columns[1].cells[5].text).to eq "puceron"
+    it 'should have tables with proper text' do
+      expect(@doc.tables[0].rows[0].cells[0].text).to eq 'ENGLISH'
+      expect(@doc.tables[0].rows[0].cells[1].text).to eq 'FRANÇAIS'
+      expect(@doc.tables[1].rows[0].cells[0].text).to eq 'Second table'
+      expect(@doc.tables[1].rows[0].cells[1].text).to eq 'Second tableau'
+      expect(@doc.tables[0].columns[0].cells[5].text).to eq 'aphids'
+      expect(@doc.tables[0].columns[1].cells[5].text).to eq 'puceron'
     end
 
-    it "should read embedded links" do
+    it 'should read embedded links' do
       expect(@doc.tables[0].columns[1].cells[1].text).to match(/^Directive/)
     end
 
@@ -109,7 +123,7 @@ describe Docx::Document do
     end
   end
 
-  describe 'editing'  do
+  describe 'editing' do
     before do
       @doc = Docx::Document.open(@fixtures_path + '/editing.docx')
     end
@@ -173,8 +187,7 @@ describe Docx::Document do
     end
 
     it 'should allow content deletion' do
-      expect{@doc.paragraphs.first.remove!}.to change{@doc.paragraphs.size}.by(-1)
-
+      expect { @doc.paragraphs.first.remove! }.to change { @doc.paragraphs.size }.by(-1)
     end
   end
 
@@ -223,7 +236,6 @@ describe Docx::Document do
     end
 
     it 'should contain a paragraph with multiple text runs' do
-
     end
 
     it 'should detect normal formatting' do
@@ -337,13 +349,11 @@ describe Docx::Document do
       temp_file.close
       temp_file.unlink
       # ensure temp file has been removed
-      expect(File.exists?(@new_doc_path)).to eq(false)
+      expect(File.exist?(@new_doc_path)).to eq(false)
     end
 
     after do
-      if File.exists?(@new_doc_path)
-        File.delete(@new_doc_path)
-      end
+      File.delete(@new_doc_path) if File.exist?(@new_doc_path)
     end
 
     context 'wps modified docx file' do
@@ -400,7 +410,7 @@ describe Docx::Document do
       expect(@doc.paragraphs[8].to_html.scan(regex).flatten.first.split(';').include?('text-align:right')).to eq(true)
     end
 
-    it "should set font size on styled paragraphs" do
+    it 'should set font size on styled paragraphs' do
       regex = /(\<p{1})[^\>]+style\=\"([^\"]+).+(<\/p>)/
       scan = @doc.paragraphs[9].to_html.scan(regex).flatten
       expect(scan.first).to eq '<p'
@@ -441,32 +451,28 @@ describe Docx::Document do
     it 'should output styled html' do
       expect(@formatted_line.to_html.scan('<span style="text-decoration:underline;"><strong><em>all</em></strong></span>').size).to eq 1
     end
-
   end
 
   describe 'replacing contents' do
     let(:replacement_file_path) { @fixtures_path + '/replacement.png' }
-    let(:temp_file_path){ Tempfile.new(['docx_gem', '.docx']).path }
-    let(:entry_path){ 'word/media/image1.png' }
-    let(:doc){ Docx::Document.open(@fixtures_path + '/replacement.docx') }
+    let(:temp_file_path) { Tempfile.new(['docx_gem', '.docx']).path }
+    let(:entry_path) { 'word/media/image1.png' }
+    let(:doc) { Docx::Document.open(@fixtures_path + '/replacement.docx') }
 
     it 'should replace existing file within the document' do
-      File.open replacement_file_path, "rb" do |io|
+      File.open replacement_file_path, 'rb' do |io|
         doc.replace_entry entry_path, io.read
       end
 
       doc.save(temp_file_path)
 
-      File.open replacement_file_path, "rb" do |io|
-        expect(Zip::File.open(temp_file_path).read entry_path).to eq io.read
+      File.open replacement_file_path, 'rb' do |io|
+        expect(Zip::File.open(temp_file_path).read(entry_path)).to eq io.read
       end
     end
 
     after do
-      if File.exists?(temp_file_path)
-        File.delete(temp_file_path)
-      end
+      File.delete(temp_file_path) if File.exist?(temp_file_path)
     end
   end
 end
-
